Add logic for diffent sub commands
