+++ README.md
@@ -1,13 +1,13 @@
 ## TaskIt
 
-It is a simple task manager that allows you to create, edit, delete and list tasks from the command line. It is built with Go.
+It is a simple task manager that allows you to create, edit, delete and list tasks from the command line. It is built with Go and Cobra.
 
 ![taskit demo gif](https://user-images.githubusercontent.com/51878265/223354705-ca2fa9c9-b054-450e-9a0b-60751c3f8ee1.gif)
 
 ### ⭐️ Features
 
 - Adding tasks
-- Deleting a task 
+- Deleting a task
 - Updating tasks
 - Completing tasks
 - Deleting all tasks
@@ -21,7 +21,7 @@ It is a simple task manager that allows you to create, edit, delete and list tas
 
 #### Local
 
-First, we need to install all the external dependencies:
+At the root og the folder run the following command to download the dependencies:
 
 ```bash
 go mod download
@@ -30,7 +30,7 @@ go mod download
 Then, we need to build the binary:
 
 ```bash
-go build ./cmd/taskit
+go build .
 ```
 
 Finally, we can run the executable:
@@ -43,23 +43,29 @@ Finally, we can run the executable:
 
 You can also use Gitpod to run this project. Gitpod is a free online IDE that allows you to run this project in the cloud. Click the button below to open this project in Gitpod:
 
-[![Open in Gitpod](https://gitpod.io/button/open-in-gitpod.svg)](https://gitpod.io/#https://github.com/Pradumnasaraf/TaskIt)
+[![Open in Gitpod](https://gitpod.io/button/open-in-gitpod.svg)](https://gitpod.io/#https://github.com/Pradumnasaraf/taskit)
 
 ### 📝 Usage
 
 We can perform tasks like add, delete, update and list tasks. We can also delete all tasks.
 
 ```
- Usage: ./task [options] [arguments]
-  Options:
-	-add "task"					     Add a new task
-	-complete [task no]				     Complete a task
-	-delete [task no]				     Delete a task
-	-update [task no] -message "message"	             Update a task (The message should be in double quotes)
-	-delete-all					     Delete all tasks
-	-list						     List all tasks
-	-help						     Show this help
-``` 
+Usage:
+  taskit [flags]
+  taskit [command]
+
+Available Commands:
+  add         Add a new task
+  delete      Delete a task
+  deleteall   Delete all tasks
+  done        Mark a task as done
+  help        Help about any command
+  list        List all tasks
+  update      Update a task by ID
+
+Flags:
+  -h, --help   help for taskit
+```
 
 ### 📜 License
