+++ cmd/root.go
@@ -3,7 +3,7 @@ package cmd
 import (
 	"os"
 
-	"github.com/Pradumnasaraf/TaskIt/cmd/task"
+	"github.com/Pradumnasaraf/taskit/cmd/task"
 	"github.com/spf13/cobra"
 )
 
@@ -19,6 +19,7 @@ var rootCmd = &cobra.Command{
 
 // Execute adds all child commands to the root command and sets flags appropriately.
 func Execute() {
+	rootCmd.CompletionOptions.DisableDefaultCmd = true
 	err := rootCmd.Execute()
 	if err != nil {
 		os.Exit(1)
@@ -26,10 +27,11 @@ func Execute() {
 }
 
 func init() {
+	// Add subcommands to root command
 	rootCmd.AddCommand(task.AddCmd)
+	rootCmd.AddCommand(task.UpdateCmd)
+	rootCmd.AddCommand(task.ListCmd)
 	rootCmd.AddCommand(task.DeleteCmd)
 	rootCmd.AddCommand(task.DeleteAllCmd)
-	rootCmd.AddCommand(task.ListCmd)
-	rootCmd.AddCommand(task.UpdateCmd)
-
+	rootCmd.AddCommand(task.DoneCmd)
 }
