+++ cmd/task/deleteAll.go
@@ -1,8 +1,9 @@
 package task
 
 import (
-	"fmt"
+	"log"
 
+	"github.com/Pradumnasaraf/taskit/handlers"
 	"github.com/spf13/cobra"
 )
 
@@ -11,6 +12,17 @@ var DeleteAllCmd = &cobra.Command{
 	Use:   "deleteall",
 	Short: "Delete all tasks",
 	Run: func(cmd *cobra.Command, args []string) {
-		fmt.Println("deleteAll called")
+
+		tasks := &handlers.Tasks{}
+		err := tasks.Load(taskFile)
+		throwErr(err)
+
+		err = tasks.DeleteAll()
+		throwErr(err)
+
+		err = tasks.Save(taskFile)
+		throwErr(err)
+
+		log.Println("All tasks deleted successfully")
 	},
 }
